Thermodynamic properties of ideal gases

Calculating $c_p, h, s$

The $c_p, h, s$ of the gas is calculated using the following NASA-9 polynomials:

\[\begin{aligned} \frac{\hat{c}_p^\circ(T)}{\overline{R}} &= && a_1T^{-2} &+ &a_2 T^{-1} &+ &a_3 &+ &a_4 T &+ &a_5T^2 &+ &a_6 T^3 &+ &a_7 T^4 &&&&\\ \frac{\hat{h}^\circ (T)}{\overline{R} T} &= &-&a_1T^{-2} &+ &a_2\frac{\ln(T)}{T} &+ &a_3 &+ &a_4\frac{T}{2} &+ &a_5\frac{T^2}{3} &+ &a_6\frac{T^3}{4} &+ &a_7\frac{T^4}{5} &+ &\frac{b_1}{T}&&\\ \frac{\hat{s}^\circ(T)}{\overline{R}} &= &-&a_1\frac{T^{-2}}{2} &- &a_2T^{-1} &+ &a_3 \ln(T) &+ &a_4 T &+ &a_5 \frac{T^2}{2} &+ &a_6 \frac{T^3}{3} &+ &a_7 \frac{T^4}{4} && &+ &b_2 \end{aligned}\]

where the $^\circ$ and $\hat{\ }$ denote standard conditions ($T_\mathrm{std} = 298.15\, \mathrm{K},\; P_\mathrm{std}=101.325\, \mathrm{kPa}$). These are dropped in the subsequent documentation for convinence.

IdealGases.Cp — Function

Cp(Tarray::AbstractVector{T}, a::AbstractVector{T}) where T

Calculates cp of the given species in J/K/mol (This is a completely non-allocating operation.)

Cp0/R = a₁T⁻² + a₂T⁻¹ + a₃ + a₄T + a₅T² + a₆T³ + a₇T⁴

Cp(T, sp::AbstractSpecies)

Calculates cp for a species type in J/K/kg.

IdealGases.h — Function

h(TT::AbstractVector{type}, a::AbstractVector{type}) where type

Calculates h of the given species in J/mol

Calcualted by:

H0/RT = -a1*T^-2 + a2*T^-1*ln(T) + a3 + a4*T/2 + a5*T^2/3 + a6*T^3/4 + a7*T^4/5 + b1/T
      = -a1*T₁   + a2*T₂*T₈      + a3 + a4*T₄/2 + a5*T₅/3  + a6*T₆/4  + a7*T₇/5  + a₈*T₂

h(T, sp::AbstractSpecies)

Calculates h for a species in J/kg

IdealGases.𝜙 — Function

𝜙(TT::AbstractVector{type},a::AbstractVector{type}) where type

Calculates the entropy complement function 𝜙=∫(cₚ/T)dT in J/K/mol

This is calculated at standard state. Tref = 298.15 K, Pref = 101325 Pa.

S0/R = -a1*T^-2/2 - a2*T^-1 + a3*ln(T) + a4*T + a5*T^2/2 + a6*T^3/3.0 + a7*T^4/4 + b2 
     = -a1*T₁/2   - a2*T₂   + a3*T₈    + a4*T₄+ a5*T₅/2  + a6*T₆/3.0  + a7*T₇/4  + a₉   

Thermodynamic derivatives

Additionally it is also useful to calculate the following derivatives,

\[\begin{aligned} \frac{d h}{d T} & = c_p\\ \frac{ds}{dT} = \frac{d\phi}{dT} & = \frac{c_p}{T}\\ \frac{d c_p}{d T} & = R_\mathrm{univ}\left(-2a_1T^{-3} - a_2 T^{-2} + a_4 + 2a_5T + 3a_6T^2 + 4a_7T^3\right)\\ \end{aligned}\]

where the last derivative $\displaystyle {\frac{dc_p}{dT}}$ is obtained by differentiating the polynomial representation of $c_p$.